requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-livereload-server==0.5.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
phonenumbers==8.13.35
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
tornado==6.4
tzdata==2024.1
